Split road_grade.rs (525 lines) into road_grade/ subdirectory
Split crates/rendering/src/road_grade.rs into focused modules:
- constants.rs: grade thresholds, colors, display constants
- helpers.rs: Bezier math, grid sampling, color interpolation
- indicators.rs: main draw_road_grade_indicators system
- tests.rs: all unit tests
- mod.rs: thin facade with re-exports
